Move graph to src
